Reorganise components
Within app/components, move each component into its own folder and
namespace them <folder_name>::Component.

This will hopefully keep this folder more manageable and allow  us to store js, css with the component.
