--- app/components/description_list_component.rb
@@ -1,13 +1,0 @@
-class DescriptionListComponent < ViewComponent::Base
-  def initialize(rows:, classes: [])
-    super
-    @rows = rows
-    @classes = classes
-  end
-
-  def classes
-    ["app-description-list"] + @classes
-  end
-
-  attr_reader :rows
-end
