+++ app/views/assessor_interface/application_forms/show.html.erb
@@ -9,7 +9,7 @@
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-full-from-desktop">
     <div class="govuk-body">
-      <%= render(ApplicationFormOverviewComponent.new(@application_form)) %>
+      <%= render(ApplicationFormOverview::Component.new(@application_form)) %>
     </div>
   </div>
 </div>
