+++ app/views/teacher_interface/identity_document/_summary.html.erb
@@ -1,4 +1,4 @@
-<%= render(CheckYourAnswersSummaryComponent.new(
+<%= render(CheckYourAnswersSummary::Component.new(
   model: application_form,
   title: I18n.t("application_form.tasks.items.identity_document"),
   fields: {
