+++ app/views/teacher_interface/registration_number/_summary.html.erb
@@ -1,4 +1,4 @@
-<%= render(CheckYourAnswersSummaryComponent.new(
+<%= render(CheckYourAnswersSummary::Component.new(
   model: application_form,
   title: I18n.t("application_form.tasks.items.registration_number"),
   fields: {
