+++ spec/components/flash_message_component_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 require "rails_helper"
 
-RSpec.describe FlashMessageComponent, type: :component do
+RSpec.describe FlashMessage::Component, type: :component do
   subject(:component) { render_inline(described_class.new(flash:)) }
 
   let(:flash) { {} }
